Move resize aware up a level * This will help performance * Should also help with the sizing of the widget, where with long names the widget won't shrink down - gets stuck
